refactor: data class model 간략화 및 적용
