--- app/src/main/java/org/sopt/dosopttemplate/data/model/responseModel/ResponseListUserSupportDto.kt
@@ -1,12 +1,0 @@
-package org.sopt.dosopttemplate.data.model.responseModel
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseListUserSupportDto(
-    @SerialName("url")
-    val url: String,
-    @SerialName("text")
-    val text: String,
-)
