+++ app/src/main/java/org/sopt/dosopttemplate/data/repository/UserRepository.kt
@@ -1,8 +1,8 @@
 package org.sopt.dosopttemplate.data.repository
 
 import org.sopt.dosopttemplate.data.datasource.UserDataSource
-import org.sopt.dosopttemplate.data.model.responseModel.toOtherUser
-import org.sopt.dosopttemplate.domain.entity.OtherUser
+import org.sopt.dosopttemplate.data.model.responseModel.toOtherUserList
+import org.sopt.dosopttemplate.domain.entity.OtherUserList
 import org.sopt.dosopttemplate.domain.repository.UserRepo
 import javax.inject.Inject
 
@@ -13,8 +13,8 @@ class UserRepository @Inject constructor(
 //        runCatching {
 //            userService.getUserList(page).data
 //        }
-    override suspend fun loadUser(page: Int): Result<List<OtherUser>> =
+    override suspend fun loadUser(page: Int): Result<OtherUserList> =
         runCatching {
-            userDataSource.getUserList(page).data.map { it.toOtherUser() }
+            userDataSource.getUserList(page).toOtherUserList()
         }
 }
