--- app/src/main/java/org/sopt/dosopttemplate/domain/entity/OtherUser.kt
@@ -1,8 +1,0 @@
-package org.sopt.dosopttemplate.domain.entity
-
-data class OtherUser(
-    val email: String,
-    val first_name: String,
-    val last_name: String,
-    val avatar: String,
-)
