+++ app/src/main/java/org/sopt/dosopttemplate/domain/repository/UserRepo.kt
@@ -1,7 +1,7 @@
 package org.sopt.dosopttemplate.domain.repository
 
-import org.sopt.dosopttemplate.domain.entity.OtherUser
+import org.sopt.dosopttemplate.domain.entity.OtherUserList
 
 interface UserRepo {
-    suspend fun loadUser(page: Int): Result<List<OtherUser>>
+    suspend fun loadUser(page: Int): Result<OtherUserList>
 }
