+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/user/UserViewHolder.kt
@@ -4,11 +4,11 @@ import androidx.recyclerview.widget.RecyclerView
 import coil.load
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.databinding.ItemUsersBinding
-import org.sopt.dosopttemplate.domain.entity.OtherUser
+import org.sopt.dosopttemplate.domain.entity.OtherUserList
 
 class UserViewHolder(private val binding: ItemUsersBinding) :
     RecyclerView.ViewHolder(binding.root) {
-    fun onBind(userData: OtherUser) =
+    fun onBind(userData: OtherUserList.OtherUser) =
         with(binding) {
             ivUserItemAvatar.load(userData.avatar) {
                 error(R.drawable.img_error)
