+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/user/UserViewModel.kt
@@ -4,17 +4,16 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
-import org.sopt.dosopttemplate.domain.entity.OtherUser
+import org.sopt.dosopttemplate.domain.entity.OtherUserList
 import org.sopt.dosopttemplate.domain.repository.UserRepo
 import javax.inject.Inject
 
 @HiltViewModel
 class UserViewModel @Inject constructor(
     private val userRepo: UserRepo,
-    // private val userRepository: UserRepository,
 ) : ViewModel() {
-    private val _loadListResult = MutableLiveData<List<OtherUser>>()
-    val loadListResult: LiveData<List<OtherUser>>
+    private val _loadListResult = MutableLiveData<List<OtherUserList.OtherUser>>()
+    val loadListResult: LiveData<List<OtherUserList.OtherUser>>
         get() = _loadListResult
 
     private val _loadListSuccess = MutableLiveData<Boolean>()
@@ -23,7 +22,7 @@ class UserViewModel @Inject constructor(
 
     suspend fun loadUserList() {
         userRepo.loadUser(2).onSuccess {
-            _loadListResult.value = it
+            _loadListResult.value = it.otherUserList
             _loadListSuccess.value = true
         }.onFailure {
             _loadListSuccess.value = false
